+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 ## [1.0.0] - 2024-05-03
 
 ### Changed 
-- `invalid_credentials` login response is now 400 instead of 404
+- `invalid_credentials` login response is now 400 instead of 404.
+- Split `transactionsRouter` into `paymentsRouter` and `depositsRouter`.
+- Split `TransactionsController` into `PaymentController` and `DepositController`.
+- Split `FinanceServices` into `PaymentServices` and `DepositServices`.
 
 ## [1.0.0] - 2024-05-02